Moved functionality to shared system
