+++ Content.Server/Devour/DevourSystem.cs
@@ -1,25 +1,15 @@
 using Content.Shared.Devour;
-using Content.Shared.DoAfter;
-using Content.Shared.Mobs.Components;
-using Content.Shared.Mobs;
 using Content.Server.Body.Systems;
 using Content.Shared.Humanoid;
 using Content.Shared.Chemistry.Components;
-using Content.Shared.Actions;
-using Robust.Shared.Containers;
-using Content.Shared.Popups;
 using Content.Server.Devour.Components;
 
 namespace Content.Server.Devour;
 
-public sealed class DevourSystem : EntitySystem
+public sealed class DevourSystem : SharedDevourSystem
 {
-    [Dependency] private readonly SharedDoAfterSystem _doAfterSystem = default!;
-    [Dependency] private readonly SharedPopupSystem _popupSystem = default!;
     [Dependency] private readonly SharedAudioSystem _audioSystem = default!;
-    [Dependency] private readonly SharedActionsSystem _actionsSystem = default!;
     [Dependency] private readonly BloodstreamSystem _bloodstreamSystem = default!;
-    [Dependency] private readonly SharedContainerSystem _containerSystem = default!;
 
     public override void Initialize()
     {
@@ -55,60 +45,5 @@ public sealed class DevourSystem : EntitySystem
             EntityManager.QueueDeleteEntity(args.Args.Target.Value);
 
         _audioSystem.PlayPvs(component.SoundDevour, uid);
-    }
-
-    private void OnStartup(EntityUid uid, DevourerComponent component, ComponentStartup args)
-    {
-        //Dragon doesn't actually chew, since he sends targets right into his stomach.
-        //I did it mom, I added ERP content into upstream. Legally!
-        component.Stomach = _containerSystem.EnsureContainer<Container>(uid, "dragon_stomach");
-
-        if (component.DevourAction != null)
-            _actionsSystem.AddAction(uid, component.DevourAction, null);
-    }
-
-    /// <summary>
-    /// The devour action
-    /// </summary>
-    private void OnDevourAction(EntityUid uid, DevourerComponent component, DevourActionEvent args)
-    {
-        if (args.Handled || component.Whitelist?.IsValid(args.Target, EntityManager) != true)
-            return;
-
-        args.Handled = true;
-        var target = args.Target;
-
-        // Structure and mob devours handled differently.
-        if (TryComp(target, out MobStateComponent? targetState))
-        {
-            switch (targetState.CurrentState)
-            {
-                case MobState.Critical:
-                case MobState.Dead:
-
-                    _doAfterSystem.TryStartDoAfter(new DoAfterArgs(uid, component.DevourTime, new DevourDoAfterEvent(), uid, target: target, used: uid)
-                    {
-                        BreakOnTargetMove = true,
-                        BreakOnUserMove = true,
-                    });
-                    break;
-                default:
-                    _popupSystem.PopupEntity(Loc.GetString("devour-action-popup-message-fail-target-alive"), uid, uid);
-                    break;
-            }
-
-            return;
-        }
-
-        _popupSystem.PopupEntity(Loc.GetString("devour-action-popup-message-structure"), uid, uid);
-
-        if (component.SoundStructureDevour != null)
-            _audioSystem.PlayPvs(component.SoundStructureDevour, uid, component.SoundStructureDevour.Params);
-
-        _doAfterSystem.TryStartDoAfter(new DoAfterArgs(uid, component.StructureDevourTime, new DevourDoAfterEvent(), uid, target: target, used: uid)
-        {
-            BreakOnTargetMove = true,
-            BreakOnUserMove = true,
-        });
     }
 }
