+++ Content.Shared/Devour/DevourActionEvent.cs
@@ -1,0 +1,5 @@
+using Content.Shared.Actions;
+
+namespace Content.Shared.Devour;
+
+public sealed class DevourActionEvent : EntityTargetActionEvent { }
